Practice Global State of react
