+++ src/App.tsx
@@ -1,4 +1,5 @@
 import './App.css';
+import { createContext, useMemo, useState, Dispatch } from 'react';
 import { Footer, Header } from './layouts';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import { Home, Examples, Pictures, Wrong, Study } from './pages';
@@ -31,25 +32,36 @@ const light_theme = {
   main_tint_more: '#c7dcf8',
 };
 
+type TestDispatch = Dispatch<React.SetStateAction<string>>;
+export const TestContext = createContext<string | null>('힝');
+export const TestDispatchContext = createContext<TestDispatch | null>(null);
+
 function App() {
+  const [globalTest, setGlobalTest] = useState('전역상태 만들어보기');
+  const testMemo = useMemo(() => globalTest, [globalTest]);
+  const dispatchMemo = useMemo(() => setGlobalTest, [setGlobalTest]);
   return (
-    <ThemeProvider theme={light_theme}>
-      <Router>
-        <div className="App">
-          <Header />
-          <Container id="container">
-            <Routes>
-              <Route path="/" element={<Home />} />
-              {/* <Route path="/examples/*" element={<Examples />} /> */}
-              <Route path="/pictures/*" element={<Pictures />} />
-              {/* <Route path="/study/*" element={<Study />} /> */}
-              <Route path="/*" element={<Wrong />} />
-            </Routes>
-            <Footer />
-          </Container>
-        </div>
-      </Router>
-    </ThemeProvider>
+    <TestContext.Provider value={testMemo}>
+      <TestDispatchContext.Provider value={dispatchMemo}>
+        <ThemeProvider theme={light_theme}>
+          <Router>
+            <div className="App">
+              <Header />
+              <Container id="container">
+                <Routes>
+                  <Route path="/" element={<Home />} />
+                  {/* <Route path="/examples/*" element={<Examples />} /> */}
+                  <Route path="/pictures/*" element={<Pictures />} />
+                  {/* <Route path="/study/*" element={<Study />} /> */}
+                  <Route path="/*" element={<Wrong />} />
+                </Routes>
+                <Footer />
+              </Container>
+            </div>
+          </Router>
+        </ThemeProvider>
+      </TestDispatchContext.Provider>
+    </TestContext.Provider>
   );
 }
